requirements.txt
awscli-local
awscli
boto3
pycrypto
SQLAlchemy
psycopg2
